+++ doc/CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added docker service
 - Added a sweet of scripts to help with managing the set up and images
 - Added github action pipeline to check changelog in PR's
+- Added World file downloads
 
 ### Fixed
 - Dockerfiles updated and docker compose aswell.